Rename lib to the same used at npm repo
