+++ test/route-beutifyLocation.js
@@ -1,4 +1,4 @@
-var PushStateTree = require('../src/pushStateTree');
+var PushStateTree = require('../src/push-state-tree');
 
 describe('PushStateTree beutifyLocation should', function() {
